+++ test/isgroebner.jl
@@ -1,12 +1,12 @@
 using Test, AbstractAlgebra, Groebner
 
 @testset "isgroebner" begin
-    R, x = polynomial_ring(GF(2^31 - 1), "x")
+    R, x = polynomial_ring(GF(Int64(2)^31 - 1), "x")
     @test Groebner.isgroebner([x])
     @test Groebner.isgroebner([x, x, x, x])
     @test !Groebner.isgroebner([x^2, x^2 + 1])
 
-    R, (x, y, z) = polynomial_ring(GF(2^31 - 1), ["x", "y", "z"], internal_ordering=:degrevlex)
+    R, (x, y, z) = polynomial_ring(GF(Int64(2)^31 - 1), ["x", "y", "z"], internal_ordering=:degrevlex)
 
     @test Groebner.isgroebner([R(1)])
     @test Groebner.isgroebner([x])
@@ -41,7 +41,7 @@ using Test, AbstractAlgebra, Groebner
 end
 
 @testset "isgroebner orderings" begin
-    R, (x, y, z) = polynomial_ring(GF(2^31 - 1), ["x", "y", "z"])
+    R, (x, y, z) = polynomial_ring(GF(Int64(2)^31 - 1), ["x", "y", "z"])
 
     @test Groebner.isgroebner([x^2 + y, y])
     @test Groebner.isgroebner([x^2 + y, y], ordering=Groebner.Lex(x, y, z))
@@ -53,7 +53,7 @@ end
 
 @testset "isgroebner certify" begin
     for certify in [false, true]
-        for field in [GF(17), GF(2^31 - 1), QQ]
+        for field in [GF(17), GF(Int64(2)^31 - 1), QQ]
             R, (x, y, z) = polynomial_ring(field, ["x", "y", "z"], internal_ordering=:degrevlex)
 
             @test Groebner.isgroebner([R(0)], certify=certify)
@@ -66,7 +66,7 @@ end
             @test !Groebner.isgroebner([x + y, x], certify=certify)
             @test Groebner.isgroebner([z * x, z * x, R(1)], certify=certify)
         end
-        fs = Groebner.Examples.rootn(3, k=GF(2^31 - 1), internal_ordering=:degrevlex)
+        fs = Groebner.Examples.rootn(3, k=GF(Int64(2)^31 - 1), internal_ordering=:degrevlex)
         @test !Groebner.isgroebner(fs, certify=certify)
         @test Groebner.isgroebner(Groebner.groebner(fs), certify=certify)
 
@@ -74,7 +74,7 @@ end
         @test !Groebner.isgroebner(fs, certify=certify)
         @test Groebner.isgroebner(Groebner.groebner(fs), certify=certify)
 
-        fs = Groebner.Examples.noonn(2, k=GF(2^31 - 1), internal_ordering=:degrevlex)
+        fs = Groebner.Examples.noonn(2, k=GF(Int64(2)^31 - 1), internal_ordering=:degrevlex)
         @test !Groebner.isgroebner(fs, certify=certify)
         @test Groebner.isgroebner(Groebner.groebner(fs), certify=certify)
 
@@ -82,7 +82,7 @@ end
         @test !Groebner.isgroebner(fs, certify=certify)
         @test Groebner.isgroebner(Groebner.groebner(fs), certify=certify)
 
-        fs = Groebner.Examples.noonn(6, k=GF(2^31 - 1), internal_ordering=:degrevlex)
+        fs = Groebner.Examples.noonn(6, k=GF(Int64(2)^31 - 1), internal_ordering=:degrevlex)
         @test !Groebner.isgroebner(fs, certify=certify)
         @test Groebner.isgroebner(Groebner.groebner(fs), certify=certify)
     end
